Created test-module for clean_text.py, corrected references in __init__.pyS, added README.md for text_manipulations package
