--- FastAPI base app/README.rst
@@ -1,25 +1,0 @@
-=============================
-Решение домашнего задания №2
-=============================
-
-Описание бизнес-задачи: преобразование и очистка произвольного текста
-
--------------------------------------------------------------------------------
-
-Установка зависимостей
--------------------------------------------------------------------------------
-- Установка Poetry, установка зависимостей из poetry.lock:
-
-.. code::
-
-    pip install poetry
-    poetry install
-
-Запуск приложения
--------------------------------------------------------------------------------
-- Выполните в командной строке:
-
-.. code::
-
-    cd FastAPI\ base\ app/
-    uvicorn app.main:app
