--- FastAPI base app/__init__.py
@@ -1,1 +1,0 @@
-from .app import text_lemmatize, clean_text, get_stopwords
