--- FastAPI base app/app/text_manipulations/README.rst
@@ -1,1 +1,0 @@
-# В этом модуле происходит автоматическая очистка и лемматизация произвольного текста
