+++ FastAPI_app/README.rst
@@ -1,0 +1,25 @@
+=============================
+Решение домашнего задания №2
+=============================
+
+Описание бизнес-задачи: очистка и лемматизация произвольного текста
+
+-------------------------------------------------------------------------------
+
+Установка зависимостей
+-------------------------------------------------------------------------------
+- Установка Poetry, установка зависимостей из poetry.lock:
+
+.. code::
+
+    pip install poetry
+    poetry install
+
+Запуск приложения
+-------------------------------------------------------------------------------
+- Выполните в командной строке:
+
+.. code::
+
+    cd FastAPI\ base\ app/
+    uvicorn app.main:app
