+++ FastAPI_app/Tests/__init__.py
@@ -1,0 +1,3 @@
+from FastAPI_app.app import *
+
+__all__ = ["text_lemmatize", "clean_text", "get_stopwords"]
