+++ FastAPI_app/Tests/test_text_manipulations/__init__.py
@@ -1,0 +1,1 @@
+from FastAPI_app.Tests import *
