+++ FastAPI_app/__init__.py
@@ -1,0 +1,1 @@
+from .app import text_lemmatize, clean_text, get_stopwords
