+++ FastAPI_app/app/text_manipulations/README.md
@@ -1,0 +1,7 @@
+Пакет очистки и лемматизации текста
+===================================
+___
+* `clean_text.py` - очищает текст от лишних символов и слов
+* `get_stopwords.py` - загружает стоп-слова из сторонних источников
+* `text_lemmatize.py` - лемматизирует произвольный текст
+* `stopwords.txt` - список стоп-слов русского языка
