+++ FastAPI_app/app/text_manipulations/__init__.py
