+++ FastAPI_app/app/text_manipulations/get_stopwords.py
