+++ FastAPI_app/app/text_manipulations/stopwords.txt
