+++ FastAPI_app/setup.cfg
@@ -1,0 +1,13 @@
+[flake8]
+count = True
+docstring-convention = google
+ignore = D100,D104,W503,PD015, I900, E800, F401
+inline-quotes = "
+max-line-length = 88
+show-source = True
+statistics = True
+per-file-ignores = tests/**.py:S101,D103,**__init__.py:F401,
+known-modules = BaseApp:[app]
+
+[isort]
+profile = black
